+++ pkg/proxy/ipvs/graceful_termination_test.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 /*
 Copyright 2019 The Kubernetes Authors.
