+++ pkg/volume/iscsi/iscsi_util_test.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 /*
 Copyright 2015 The Kubernetes Authors.
